+++ CosmosDBShell/Azure.Data.Cosmos.Shell.Commands/ReplaceCommand.cs
@@ -54,15 +54,16 @@ internal class ReplaceCommand : CosmosCommand
 
         var partitionKeyPaths = await CosmosResourceFacade.GetPartitionKeyPathsAsync(connectedState, databaseName!, containerName!, token);
 
-        await ReplaceItemsAsync(container, partitionKeyPaths, jsonOpt, this.ETag, token);
+        var totalCharge = await ReplaceItemsAsync(container, partitionKeyPaths, jsonOpt, this.ETag, token);
 
         return new CommandState
         {
             Result = new ShellJson(SuccessDocument.RootElement.Clone()),
+            RequestCharge = totalCharge,
         };
     }
 
-    private static async Task ReplaceItemsAsync(Container container, IReadOnlyList<string> partitionKeyPaths, string jsonInput, string? etag, CancellationToken token)
+    private static async Task<double> ReplaceItemsAsync(Container container, IReadOnlyList<string> partitionKeyPaths, string jsonInput, string? etag, CancellationToken token)
     {
         try
         {
@@ -76,11 +77,10 @@ internal class ReplaceCommand : CosmosCommand
                     throw new CommandException("replace", MessageService.GetString("command-replace-error-etag_array_not_supported"));
                 }
 
-                await ReplaceArrayAsync(container, partitionKeyPaths, root, token);
-                return;
+                return await ReplaceArrayAsync(container, partitionKeyPaths, root, token);
             }
 
-            await ReplaceOneAsync(container, partitionKeyPaths, root, etag, token, printSuccess: true);
+            return await ReplaceOneAsync(container, partitionKeyPaths, root, etag, token, printSuccess: true);
         }
         catch (JsonException ex)
         {
@@ -88,7 +88,7 @@ internal class ReplaceCommand : CosmosCommand
         }
     }
 
-    private static async Task ReplaceArrayAsync(Container container, IReadOnlyList<string> partitionKeyPaths, JsonElement arrayRoot, CancellationToken token)
+    private static async Task<double> ReplaceArrayAsync(Container container, IReadOnlyList<string> partitionKeyPaths, JsonElement arrayRoot, CancellationToken token)
     {
         int successCount = 0;
         int failCount = 0;
@@ -133,6 +133,8 @@ internal class ReplaceCommand : CosmosCommand
                     "total",
                     successCount + failCount));
         }
+
+        return charge;
     }
 
     private static async Task<double> ReplaceOneAsync(Container container, IReadOnlyList<string> partitionKeyPaths, JsonElement item, string? etag, CancellationToken token, bool printSuccess)